fixing stats double free error
Former-commit-id: a5a2c616c0ae6d2ffaa19da35da6d4b9cd5ca72f [formerly 68a2c8b8f34f25debd3b38b293af4710fd14a805] [formerly e76df6b732fc6a99123ccd2d85b3c81e109213e1 [formerly 49216599683c5cc0ed9be8ae33cac9709aac530f]]
Former-commit-id: c437fdbafc704846010bbab8b48e5daeb6dcee5f [formerly 4148cc94adca6ecc82dd277d3e88d67273a6ccfa]
Former-commit-id: e7cf8e2e62f4f26161035dfb20354dabfe427b39

+++ src/util/utils.h
@@ -219,7 +219,7 @@ namespace FT{
             
             void resize(int new_size);
             
-            void update(float index,
+            void update(int index,
                         float timer_count,
                         float bst_score,
                         float bst_score_v,
